Updated all comments

+++ WordFinder/src/main.cpp
@@ -1,20 +1,20 @@
 /* 
  * Author: Cameron Cipriano
- * Date: 12/28/2018
+ * Date: 1/12/2019
  * Description:
- *
+ *      Main runner for the BoggleSolver. This is where the user can enter all of the necessary details
+ *      to generate a new intance of the BoggleGame. The user has the ability to change everything,
+ *      including which dictionary text file to use should there be differences in them.
  */
 
 #include <cstdlib>
 #include <string>
 #include <iostream>
-
-// My headers 
 #include "../headers/BoggleGame.h"
 
 int main(int argc, char** argv) {
     
-    // for an instance of the BoggleGame, we need the board size, minWordLen, and the letters
+    // for an instance of the BoggleGame, we need the dictionary text file path, board size, minWordLen, and the letters
     size_t boardSize = 0;
     size_t propNumLetters = 0;
     size_t numLetters;
@@ -22,7 +22,7 @@ int main(int argc, char** argv) {
     string boardLetters;
     string dictPath;
 
-    switch(argc) {
+    switch(argc) { // error handling for not enough arguments of different kinds
         case 1:
             cerr << "\nNot enough arguments!" << endl;
             cerr << "<filename> <dictFile> <boardSize> <minWordLen> <letters...>\n" << endl;
@@ -38,14 +38,14 @@ int main(int argc, char** argv) {
             cerr << "<filename> <dictFile> <boardSize> <minWordLen> <letters...>\n" << endl;
             exit(3);
         case 4:
-            boardSize = stoi(argv[2]);
+            boardSize = stoul(argv[2]);
             propNumLetters = (int)(boardSize * boardSize);
             cerr << "\nNot enough arguments!" << endl;
             cerr << "Please type in the board letters! Based on your input, you need ";
             cerr << propNumLetters << endl;
             cerr << "<filename> <dictFile> <boardSize> <minWordLen> <letters...>\n" << endl;
             exit(4);            
-        default:
+        default: // this means that sufficient/more arguments have been provided
             dictPath = argv[1];
             boardSize = stoul(argv[2]);
             propNumLetters = (int)(boardSize * boardSize);
@@ -53,7 +53,7 @@ int main(int argc, char** argv) {
             boardLetters = argv[4];
             numLetters = boardLetters.length();
 
-            if (numLetters != propNumLetters) {
+            if (numLetters != propNumLetters) { // enough arguments have been provided, but does not fit the needed amount total
                 cerr << "\nNot enough letters!" << endl;
                 cerr << "You entered " << numLetters << " but you still need " << (propNumLetters - numLetters) << " more!\n" << endl;
                 exit (5);
